+++ Gemfile
@@ -8,3 +8,4 @@ gem 'mailgun-ruby', '~>1.1.6'
 gem 'mail', '~>2.6.6'
 gem 'mailchimp-api', '~>2.0.4'
 gem 'titleize', '~>1.4.1'
+gem 'pry-byebug', '~>3.6.0'
